add some conditions when data is empty

+++ controllers/penduduk.controller.js
@@ -244,15 +244,17 @@ exports.deletePendudukPadaKK = async (req, res) => {
         nomor_kartu_keluarga: findKK.no_kk,
       },
       async (err, res) => {
-        await PendudukKeluarSchema.findByIdAndUpdate(
-          { _id: res._id },
-          {
-            $pull: { penduduk_keluar_desa: req.params.id_penduduk },
-          }
-        );
+        if (res !== null) {
+          await PendudukKeluarSchema.findByIdAndUpdate(
+            { _id: res._id },
+            {
+              $pull: { penduduk_keluar_desa: req.params.id_penduduk },
+            }
+          );
 
-        if (res.penduduk_keluar_desa.length === 0) {
-          await PendudukKeluarSchema.deleteOne({ _id: res._id });
+          if (res.penduduk_keluar_desa.length === 0) {
+            await PendudukKeluarSchema.deleteOne({ _id: res._id });
+          }
         }
       }
     );
@@ -269,9 +271,10 @@ exports.deletePendudukPadaKK = async (req, res) => {
           nomor_kartu_keluarga: findKK.no_kk,
         },
         async (err, res) => {
-          if (res.penduduk_keluar_desa.length === 0) {
-            await PendudukKeluarSchema.deleteOne({ _id: res._id });
-          }
+          if (res !== null)
+            if (res.penduduk_keluar_desa.length === 0) {
+              await PendudukKeluarSchema.deleteOne({ _id: res._id });
+            }
         }
       );
